+++ extension.js
@@ -4,6 +4,7 @@ const vscode = require('vscode');
 var orgDeployRetrieve = require('./src/orgDeployRetrieve');
 var orgWorkspaceColor = require('./src/orgWorkspaceColor');
 var lwcHtmlMarkup = require('./src/lwcHtmlMarkup');
+var codeDiagnostics = require('./src/codeDiagnostics');
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -33,17 +34,35 @@ function activate(context) {
 		orgWorkspaceColor.changeWorkspaceColorToDefault();
     });
 
-	// reset workspace color
+	// todo: creating webview 
     let insertLightningMarkup = vscode.commands.registerCommand('sfdevtools.insertLightningMarkup', function () {
 		lwcHtmlMarkup.selectMarkup();
     });
+
+	// Open lwc component library
+    let openLwcLibrary = vscode.commands.registerCommand('sfdevtools.openLwcLibrary', function () {
+		orgDeployRetrieve.openLwcLibrary();
+    });
+
+	// todo: create diagnostics  
+	const collection = vscode.languages.createDiagnosticCollection('test');
+
+	if (vscode.window.activeTextEditor) {
+		codeDiagnostics.upsertDiagnostics(vscode.window.activeTextEditor.document, collection);
+	}
+	context.subscriptions.push(vscode.window.onDidChangeActiveTextEditor(editor => {
+		if (editor) {
+			codeDiagnostics.upsertDiagnostics(editor.document, collection);
+		}
+	}));
 
 	context.subscriptions.push(
 		deployDisposable, 
 		retrieveDisposable, 
 		changeWorkspaceColor, 
 		changeWorkspaceColorToDefault,
-		insertLightningMarkup
+		insertLightningMarkup,
+		openLwcLibrary
 	);
 }